add to playlist rather than play directly
Tiny tweak - we now add to a playlist with a known ID, rather than play the track directly. when we play the track directly, the Spotify player will just stop after the track completes. when we instead add to a playlist (and then start playback of the last track in that playlist), we keep jammin to infinite Spotify recs after the song is done